updated prefabs + collectable script
updated the player prefab to ensure the collectable system is working properly. also made changes to the collectable script, it seems to have been attempting to trigger on collision instead of a button press. Collect collectables with F, close popups with E.

also includes the almost completed UI system for showing which collectables a player has missed