Finished settings menu stuff
menus are all properly linked now, deleting all of the buttons and re-adding them seems to have fixed it, and a lot of menu assets have been resized to look more cohesive. I think there is a bug with the resetting button not functioning after resetting, but I think I know how to fix this so it'll be dealt with soon. I've also updated the menu prefabs so it should be functional wherever the menus already exist